feat: 메인 페이지 디렉토리 변경
