+++ src/app/[locale]/__container.tsx
