+++ src/app/[locale]/page.tsx
@@ -1,5 +1,39 @@
-import { redirect } from 'next/navigation';
+import { backgroundPhotoQueryOptions, type PhotoParams } from '@/entities/background';
+import { bibleTodayQueryOptions } from '@/features/bible';
+import { newsQueryOptions } from '@/features/news';
+import { PWAInstallPrompt, type RouteProps } from '@/shared';
+import { dehydrate, HydrationBoundary, QueryClient } from '@tanstack/react-query';
+import Container from './__container';
 
-export default async function LocalePage() {
-  redirect('/home');
-}
+const HomePage = async ({ params }: RouteProps) => {
+  const PHOTO_LIST_SIZE = 10;
+  const { locale } = await params;
+  const backgroundPhotoParams = {
+    query: 'river natural empty center',
+    page: 1,
+    perPage: PHOTO_LIST_SIZE,
+    orientation: 'landscape'
+  } satisfies PhotoParams;
+
+  const queryClient = new QueryClient();
+  await Promise.all([
+    queryClient.prefetchQuery(bibleTodayQueryOptions(locale)),
+    queryClient.prefetchQuery(newsQueryOptions(locale)),
+    queryClient.prefetchQuery(backgroundPhotoQueryOptions(backgroundPhotoParams))
+  ]);
+  const dehydratedState = dehydrate(queryClient);
+
+  const randomBackgroundPhotoIndex = new Date().getUTCDate() % PHOTO_LIST_SIZE;
+
+  return (
+    <HydrationBoundary state={dehydratedState}>
+      <Container
+        backgroundPhotoParams={backgroundPhotoParams}
+        randomBackgroundPhotoIndex={randomBackgroundPhotoIndex}
+      />
+      <PWAInstallPrompt />
+    </HydrationBoundary>
+  );
+};
+
+export default HomePage;
